# Conflicts:
#	go.mod
#	go.sum

+++ cmd/sfnear/cli/node.go
@@ -28,14 +28,14 @@ import (
 func registerCommonNodeFlags(cmd *cobra.Command, flagPrefix string, managerAPIAddr string) {
 	cmd.Flags().String(flagPrefix+"path", "neard", "Command that will be launched by the node manager")
 	cmd.Flags().String(flagPrefix+"data-dir", "{sf-data-dir}/{node-role}/data", "Directory for node data ({node-role} is either mindreader, peering or dev-miner)")
-	cmd.Flags().String(flagPrefix+"config-file", "./{node-role}/config.json", "Node configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder")
-	cmd.Flags().String(flagPrefix+"genesis-file", "./{node-role}/genesis.json", "Node configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder")
-	cmd.Flags().String(flagPrefix+"node-key-file", "./{node-role}/node_key.json", "Node key configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder")
+	cmd.Flags().String(flagPrefix+"config-file", "./{node-role}/config.json", "Node configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder Use {hostname} label to use short hostname in path")
+	cmd.Flags().String(flagPrefix+"genesis-file", "./{node-role}/genesis.json", "Node configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder. Use {hostname} label to use short hostname in path")
+	cmd.Flags().String(flagPrefix+"node-key-file", "./{node-role}/node_key.json", "Node key configuration file where ({node-role} is either mindreader, peering or dev-miner), the file is copied inside the {sf-data-dir}/{node-role}/data folder. Use {hostname} label to use with short hostname in path")
 	cmd.Flags().Bool(flagPrefix+"debug-deep-mind", false, "[DEV] Prints deep mind instrumentation logs to standard output, should be use for debugging purposes only")
 	cmd.Flags().Bool(flagPrefix+"log-to-zap", true, "Enable all node logs to transit into node's logger directly, when false, prints node logs directly to stdout")
 	cmd.Flags().String(flagPrefix+"manager-api-addr", managerAPIAddr, "Near node manager API address")
 	cmd.Flags().Duration(flagPrefix+"readiness-max-latency", 30*time.Second, "Determine the maximum head block latency at which the instance will be determined healthy. Some chains have more regular block production than others.")
-	cmd.Flags().String(flagPrefix+"node-extra-arguments", "", "Extra arguments to be passed when executing superviser binary")
+	cmd.Flags().String(flagPrefix+"arguments", "", "If not empty, overrides the list of default node arguments (computed from node type and role). Start with '+' to append to default args instead of replacing. ")
 
 	// FIXME: Right now our near-dm-indexer doesn't support it, we should plan on adding it!
 	// cmd.Flags().String(flagPrefix+"node-boot-nodes", "", "Set the node's boot nodes to bootstrap network from")
@@ -85,25 +85,28 @@ func nodeFactoryFunc(flagPrefix, kind string, appLogger, nodeLogger **zap.Logger
 		configFile := replaceNodeRole(kind, viper.GetString(flagPrefix+"config-file"))
 		genesisFile := replaceNodeRole(kind, viper.GetString(flagPrefix+"genesis-file"))
 		nodeKeyFile := replaceNodeRole(kind, viper.GetString(flagPrefix+"node-key-file"))
+
+		hostname, _ := os.Hostname()
+		configFile = replaceHostname(hostname, configFile)
+		genesisFile = replaceHostname(hostname, genesisFile)
+		nodeKeyFile = replaceHostname(hostname, nodeKeyFile)
+
 		readinessMaxLatency := viper.GetDuration(flagPrefix + "readiness-max-latency")
 		debugDeepMind := viper.GetBool(flagPrefix + "debug-deep-mind")
 		logToZap := viper.GetBool(flagPrefix + "log-to-zap")
 		shutdownDelay := viper.GetDuration("common-system-shutdown-signal-delay") // we reuse this global value
 		httpAddr := viper.GetString(flagPrefix + "manager-api-addr")
 
+		arguments := viper.GetString(flagPrefix + "arguments")
 		nodeArguments, err := buildNodeArguments(
 			nodeDataDir,
 			flagPrefix,
 			kind,
+			arguments,
 		)
 		if err != nil {
 			return nil, fmt.Errorf("cannot build node bootstrap arguments")
 		}
-		extraArgs := getExtraArguments(flagPrefix)
-		if len(extraArgs) > 0 {
-			nodeArguments = append(nodeArguments, extraArgs...)
-		}
-
 		metricsAndReadinessManager := buildMetricsAndReadinessManager(flagPrefix, readinessMaxLatency)
 
 		superviser := nodemanager.NewSuperviser(
@@ -221,16 +224,32 @@ func (b *bootstrapper) Bootstrap() error {
 		return fmt.Errorf("create all dirs of %q: %w", b.nodeDataDir, err)
 	}
 
-	if err := copyFile(ctx, b.configFile, configFileInDataDir); err != nil {
-		return fmt.Errorf("unable to copy config file %q to %q: %w", b.configFile, configFileInDataDir, err)
+	exists, err := fileExists(configFileInDataDir)
+	if err != nil {
+		return err
 	}
-
-	if err := copyFile(ctx, b.genesisFile, genesisFileInDataDir); err != nil {
-		return fmt.Errorf("unable to copy genesis file %q to %q: %w", b.genesisFile, genesisFileInDataDir, err)
+	if !exists {
+		if err := copyFile(ctx, b.configFile, configFileInDataDir); err != nil {
+			return fmt.Errorf("unable to copy config file %q to %q: %w", b.configFile, configFileInDataDir, err)
+		}
 	}
-
-	if err := copyFile(ctx, b.nodeKeyFile, nodeKeyFileInDataDir); err != nil {
-		return fmt.Errorf("unable to copy node key file %q to %q: %w", b.nodeKeyFile, nodeKeyFileInDataDir, err)
+	exists, err = fileExists(genesisFileInDataDir)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		if err := copyFile(ctx, b.genesisFile, genesisFileInDataDir); err != nil {
+			return fmt.Errorf("unable to copy genesis file %q to %q: %w", b.genesisFile, genesisFileInDataDir, err)
+		}
+	}
+	exists, err = fileExists(nodeKeyFileInDataDir)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		if err := copyFile(ctx, b.nodeKeyFile, nodeKeyFileInDataDir); err != nil {
+			return fmt.Errorf("unable to copy node key file %q to %q: %w", b.nodeKeyFile, nodeKeyFileInDataDir, err)
+		}
 	}
 
 	return nil
@@ -238,24 +257,35 @@ func (b *bootstrapper) Bootstrap() error {
 
 type nodeArgsByRole map[string]string
 
-func buildNodeArguments(nodeDataDir, flagPrefix, nodeRole string) ([]string, error) {
+func buildNodeArguments(nodeDataDir, flagPrefix, nodeRole string, args string) ([]string, error) {
 	typeRoles := nodeArgsByRole{
-		"peering":    "--home={node-data-dir} run",
-		"mindreader": "--home={node-data-dir} run",
+		"peering":    "--home={node-data-dir} {extra-arg} run",
+		"mindreader": "--home={node-data-dir} {extra-arg} run",
 	}
 
-	roleArgs, ok := typeRoles[nodeRole]
+	argsString, ok := typeRoles[nodeRole]
 	if !ok {
 		return nil, fmt.Errorf("invalid node role: %s", nodeRole)
 	}
-	args := strings.Fields(strings.Replace(roleArgs, "{node-data-dir}", nodeDataDir, -1))
+
+	if strings.HasPrefix(args, "+") {
+		argsString = strings.Replace(argsString, "{extra-arg}", args[1:], -1)
+	} else if args == "" {
+		argsString = strings.Replace(argsString, "{extra-arg}", "", -1)
+	} else {
+		argsString = args
+	}
+
+	argsString = strings.Replace(argsString, "{node-data-dir}", nodeDataDir, -1)
+
+	argsSlice := strings.Fields(argsString)
 
 	bootNodes := viper.GetString(flagPrefix + "node-boot-nodes")
 	if bootNodes != "" {
-		args = append(args, "--boot-nodes", viper.GetString(flagPrefix+"node-boot-nodes"))
+		argsSlice = append(argsSlice, "--boot-nodes", viper.GetString(flagPrefix+"node-boot-nodes"))
 	}
 
-	return args, nil
+	return argsSlice, nil
 }
 
 func buildMetricsAndReadinessManager(name string, maxLatency time.Duration) *nodeManager.MetricsAndReadinessManager {
@@ -270,11 +300,10 @@ func buildMetricsAndReadinessManager(name string, maxLatency time.Duration) *nod
 	return metricsAndReadinessManager
 }
 
-func getExtraArguments(prefix string) (out []string) {
-	extraArguments := viper.GetString(prefix + "-node-extra-arguments")
-	if extraArguments != "" {
-		out = append(out, strings.Split(extraArguments, " ")...)
-	}
+func replaceNodeRole(nodeRole, in string) string {
+	return strings.Replace(in, "{node-role}", nodeRole, -1)
+}
 
-	return
+func replaceHostname(hostname, in string) string {
+	return strings.Replace(in, "{hostname}", hostname, -1)
 }
